requirements.txt
# Dependencias principales
agno>=0.1.0
groq>=0.4.0
httpx>=0.25.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Para el agente de marketing
langchain-apify>=0.1.0
apify-client>=1.6.0
requests>=2.31.0

# Base de datos
aiosqlite>=0.19.0

# Opcionales para desarrollo
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Variables de entorno
python-dotenv>=1.0.0